+++ src/config-firebase.js
@@ -1,21 +1,23 @@
-// Import the functions you need from the SDKs you need
-import { initializeApp } from "firebase/app";
-import { getAnalytics } from "firebase/analytics";
-// TODO: Add SDKs for Firebase products that you want to use
-// https://firebase.google.com/docs/web/setup#available-libraries
+// // Import the functions you need from the SDKs you need
+// import { initializeApp } from "firebase/app";
+// import { getAnalytics } from "firebase/analytics";
+// import 'firebase/auth';
 
-// Your web app's Firebase configuration
-// For Firebase JS SDK v7.20.0 and later, measurementId is optional
-const firebaseConfig = {
-  apiKey: "AIzaSyAbMAlU7tKoqM8G0vzW63Je5YM-gqIA-MI",
-  authDomain: "fit-sync-8974c.firebaseapp.com",
-  projectId: "fit-sync-8974c",
-  storageBucket: "fit-sync-8974c.appspot.com",
-  messagingSenderId: "1069101892727",
-  appId: "1:1069101892727:web:070ac7c457c1e444e83272",
-  measurementId: "G-T7HF2BQQSR"
-};
+// // TODO: Add SDKs for Firebase products that you want to use
+// // https://firebase.google.com/docs/web/setup#available-libraries
 
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
-const analytics = getAnalytics(app);
+// // Your web app's Firebase configuration
+// // For Firebase JS SDK v7.20.0 and later, measurementId is optional
+// const firebaseConfig = {
+//   apiKey: "AIzaSyAbMAlU7tKoqM8G0vzW63Je5YM-gqIA-MI",
+//   authDomain: "fit-sync-8974c.firebaseapp.com",
+//   projectId: "fit-sync-8974c",
+//   storageBucket: "fit-sync-8974c.appspot.com",
+//   messagingSenderId: "1069101892727",
+//   appId: "1:1069101892727:web:070ac7c457c1e444e83272",
+//   measurementId: "G-T7HF2BQQSR"
+// };
+
+// // Initialize Firebase
+// const app = initializeApp(firebaseConfig);
+// const analytics = getAnalytics(app);
